main_users_db/pyproject.toml
[tool.poetry]
name = "main-users-db"
version = "0.1.0"
description = "-"
authors = ["Vitaly <lildid32@mail.ru>"]
readme = "README.md"

packages = [{include="main"}]
package-mode = true


[tool.poetry.scripts]
run-server = "main.server:start"

[tool.pytest.ini_options]
asyncio_mode = "auto"

pythonpath = [
    "."
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
uvicorn = "^0.32.0"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
pydantic-extra-types = "^2.9.0"
pytest-asyncio = "^0.24.0"
phonenumbers = "^8.13.49"
passlib = "^1.7.4"
pydantic-settings = "^2.6.1"
redis = "^5.2.0"
fastapi-utils = "^0.8.0"
typing-inspect = "^0.9.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
httpx = "^0.28.1"
alembic-postgresql-enum = "^1.4.0"
gunicorn = "^23.0.0"
pika = "^1.3.2"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
